Update RegExodus dependency (gwt.xml change).
This one is important for gdx-liftoff, which tries to pick the most recent version for any dependency (0.1.15 for RegExodus). That most recent RegExodus uses a different `inherits` line in its .gwt.xml file, because I discovered with many of my libraries that using the root folder (`regexodus`) causes problems that are fixed by using a subfolder (`regexodus.regexodus`).