Fix #28 .
The fix is relatively simple; if glyph.color was null, then its color wouldn't be set, including its alpha. Now we just check for alpha being less than 1 if glyph.color is null, and use our already-configured tint color in that case without needing to use glyph.color. The only tricky part is that I needed to avoid modifying color so it could be used when glyph.color is null... So setColors is given a packed float color assembled from color and glyph.color channels, and is otherwise the same.